Moved rate map provider to providers
