+++ currency/src/main/java/com/mineraltree/currency/providers/FixerProvider.java
@@ -4,7 +4,6 @@ import akka.http.javadsl.model.Query;
 import akka.japi.Pair;
 import com.fasterxml.jackson.databind.node.BooleanNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.mineraltree.currency.RateMapProvider;
 import com.mineraltree.currency.dto.CurrencyRates.Builder;
 import com.mineraltree.http.RequestUtils;
 import com.typesafe.config.Config;
